vmm: config: set default socket value for virtio-fs in api code path
Now that we have a native virtio-fs, socket parameter is not a must.
We should give it a default serde value when input is the restful api.
(There is already a default value for the normal comamnd way, so it's
ok) Also, do a check to avoid socket and backendfs_config exist in the
same time.

Signed-off-by: Hao Xu <howeyxu@tencent.com>